refactor: break up settings js into smaller modules
